+++ src/cli/extensions/timer.ts
@@ -11,6 +11,12 @@ export default (toolbox: GluegunToolbox) => {
           const elapsed = (timer() / 1000).toFixed(2);
           const time = print.colors.muted(`[${elapsed} s]`);
           print.success(`${msg} ${time}\n`);
+        },
+        fail(msg = 'Error!') {
+          const elapsed = (timer() / 1000).toFixed(2);
+          const time = print.colors.muted(`[${elapsed} s]`);
+          print.error(`${msg} ${time}\n`);
+          process.exit(1);
         }
       }
     }